Freshen up test
